+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add Group methods to get subgroups and supergroups
 - limit path column being inserted to size of column
 - add top level db function
+- add request environment variables to admin environment view
 
 ## [6.21.0] - 2023-05-01
 - add quiet logging completion of ajax requests